Solution

+++ services/movie_session.py
@@ -41,7 +41,8 @@ def update_movie_session(
 
 
 def delete_movie_session_by_id(session_id: int) -> None:
-    MovieSession.objects.get(id=session_id).delete()
+    if MovieSession.objects.filter(id=session_id).exists():
+        MovieSession.objects.get(id=session_id).delete()
 
 
 def get_taken_seats(movie_session_id: int) -> list[dict]:
